added exceptions for item

+++ src/main/java/com/example/practice/service/ItemService.java
@@ -2,6 +2,8 @@ package com.example.practice.service;
 
 import com.example.practice.dto.ItemDto;
 import com.example.practice.entity.Item;
+import com.example.practice.exceptions.ItemExistsException;
+import com.example.practice.exceptions.ItemNotFoundException;
 import com.example.practice.mapper.ItemMapper;
 import com.example.practice.repository.ItemRepository;
 import lombok.RequiredArgsConstructor;
@@ -17,30 +19,35 @@ public class ItemService {
     private final ItemRepository itemRepository;
     private final ItemMapper itemMapper;
 
+    public void create(Item item) {
+        if (itemRepository.existsById(item.getId())) {
+            throw new ItemExistsException("Item with id = " + item.getId() + "already exists");
+        }
+        itemRepository.save(item);
+    }
+
     public Optional<Item> read(Long id) {
+        if (!itemRepository.existsById(id)) {
+            throw new ItemNotFoundException();
+        }
         return itemRepository.findById((id));
     }
 
-    public void create(Item item) {
+    public void update(Item item) {
+        if (!itemRepository.existsById(item.getId())) {
+            throw new ItemNotFoundException();
+        }
         itemRepository.save(item);
     }
 
     public void delete(Long id) {
+        if (!itemRepository.existsById(id)) {
+            throw new ItemNotFoundException();
+        }
         itemRepository.deleteById(id);
     }
 
     public List<ItemDto> readAll() {
-        var t = itemRepository.findAll();
-        var mt = itemMapper.toDto(t);
-        return mt;
+        return itemMapper.toDto(itemRepository.findAll());
     }
-
-    public boolean update(Item item) {
-        if (itemRepository.existsById(item.getId())) {
-            itemRepository.save(item);
-            return true;
-        }
-        return false;
-    }
-
 }
